tests/wpt/web-platform-tests/referrer-policy/css-integration/font-face/external-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - Font from imported stylesheet (external)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="never">
  </head>
  <body>
    <p>Check that resources from imported stylesheets (loaded from external
    stylesheets) are loaded with the referrer and referrer policy from the
    external stylesheet.</p>

    <div class="styled"></div>

    <script>
      let css_test = async_test("Font from imported stylesheet (external).");
      let id = token();
      let css_url = location.protocol + "//www1." + location.hostname + ":" +
        location.port +
        "/referrer-policy/generic/subresource/stylesheet.py?id=" + id +
        "&import-rule" + "&type=font";
      let url_prefix = location.protocol + "//" + location.hostname + ":" + location.port;
      let css_referrer = url_prefix +
        "/referrer-policy/generic/subresource/stylesheet.py?id=" + id + "&type=font";
      let font_url = url_prefix + "/referrer-policy/generic/subresource/font.py" +
                     "?id=" + id + "&report-headers" + "&type=font";

      let link = document.createElement("link");
      link.href = css_url;
      link.rel = "stylesheet";
      link.onload = function() {
        css_test.step_timeout(
            queryXhr.bind(this, font_url,
                          function(message) {
                            css_test.step(function() {
                              assert_own_property(message, "headers");
                              assert_own_property(message, "referrer");
                              assert_equals(message.referrer, css_referrer);
                            });
                            css_test.done();
                          }),
            1000);
      };
      document.head.appendChild(link);
    </script>

    <div id="log"></div>
  </body>
</html>
